refactor: fix Ryu backend, increase precision limits, and improve ARM toolchain
- Ryu: fix buffer reservations and formatting logic by using type-specific
  MAX_10_EXP and implementing trailing zero stripping.
- Core: increase maximum width and precision limits from 254 to 65534
  by switching from uint8_t to uint16_t.
- Build: modernize ARM toolchain detection, improve sysroot handling,
  and update CMake presets for better QEMU integration.
- Backend: switch default float backend to STRFROM for better reliability.
- Misc: improve version detection when git is missing and remove VLAs from tests.

+++ src/yio/private/yio_float_strfrom_printf.c
@@ -33,7 +33,7 @@ void YYIO_create_format_string_generic(char *restrict fmt, size_t fmtsize,
 	*fmtpnt++ = '%';
 	if (precision0 != 0) {
 		*fmtpnt++ = '.';
-		const size_t precision = yio_precision_get_default((uint8_t)precision0, 0);
+		const size_t precision = yio_precision_get_default(precision0, 0);
 		const int len = yio_snstream(fmtpnt, INT_MAX, precision);
 		(void)len;
 		assert(len > 0);
